requirements.txt
flask==2.3.3
gunicorn==21.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.3.0
scikit-learn==1.7.2
joblib==1.4.2
tensorflow-cpu==2.15.0